requirements.txt
streamlit==1.37.1
openai-whisper==20231117
torch==2.2.2
transformers==4.41.2
pydub==0.25.1
ffmpeg-python==0.2.0
reportlab==4.2.0
python-docx==1.1.2
tqdm==4.66.4
nltk==3.9.1
scikit-learn==1.3.2
pandas==2.2.3
numpy==1.26.4